Clean up "gulp presubmit" errors. Rename ads/google/adsense.js and utils.js to a4a-adsense.js and a4a-utils.js, then revert ads/google/adsense.js and utils.js to tunk (a4a-core) versions.
This works, and passes our tests. It still needs to be cleaned up, but will allow
work on a4a doubleclick to proceed.
